+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Better keyboard support on JS-powered links (delete/cancel upload buttons)
   - rather than make the `keypress` event trigger the `click` event, just set an `href` and `click` magically works
 - Properly pluralize human durations (if 1 of the unit)
+- Return an error when cleaning up an upload with unknown metadata version instead of deleting it
 
 ## [v0.2.1] - 2022-01-21